refactor(input): rename
